Multi-Agent color coded video making
Creating a new environment called multiagent_ngsim_env_videomaker.jl to
enable the creation of multiagent color coded videos. The reason that
this cannot be done using the existing multiagent_ngsim_env.jl file is
that the added features that track the list of colliding, offroad and
hardbraking vehicles which are passed into the rendering were causing
the training process in imitate.py to break.
A line was added in AutoEnvs.jl, make.jl and utils.py to recognize this
new environment.
multiAgentMakeVideo.py is the file that actually creates the video.
Specify the experiment details within so that it can load the policy.
Run it using python multiAgentMakeVideo.py